requirements.txt
openai
gradio
pillow
